イラストモードの実装

+++ index.py
@@ -7,7 +7,7 @@ import uuid
 import json
 import promptOptimize
 import requests,base64
-
+from io import BytesIO
 # Discord.pyからクライアントインスタンスを作成するよおおおおおおおおおおお
 client = discord.Client(intents = discord.Intents.all())
 intents = discord.Intents.default()
@@ -45,6 +45,7 @@ async def on_message(message:discord.Message):
         return
 
     if message.author != client and (f"<@{APPLICATION_ID}>" in message.content) or not message.guild:
+        mode = 0
         json_load = json.load(open('permission.json','r'))
         # permission.jsonを参照し、一般ユーザーの使用を許可するかどうか確認し、許可しない場合はreturn(メンテナンス中など、生成してほしくない時に使用する)
         if not json_load['normal_user'] and not message.author.name == 'yomi4486' :
@@ -57,12 +58,13 @@ async def on_message(message:discord.Message):
         # 送信されたメッセージがメンションのみだった場合は処理を行わず、Botの説明を軽く行う
         if not JPprompt.replace(' ', '').replace('　', '') == '':
             if len(JPprompt) > 499:
-                await message.channel.send('単語が長すぎるよ！500文字未満にしてね！')
+                await message.channel.send('プロンプトが長すぎるよ！500文字未満にしてね！')
                 return
             
             if 'イラスト' in JPprompt or 'illust' in JPprompt.lower() or 'アニメ' in JPprompt or 'anime' in JPprompt.lower():
                 JPprompt = JPprompt.lower().replace('illust','').replace('anime','').replace('アニメ','').replace('イラスト','')
                 JPprompt = JPprompt + ',anime,waifu:1.2'
+                mode = 1
             
             JPprompt = promptOptimize.optimize(JPprompt)
             # 正確性が増すので英語に翻訳してからプロンプトに突っ込む
@@ -77,27 +79,34 @@ async def on_message(message:discord.Message):
                 while not prompt == replaceprompt:
                     replaceprompt = prompt
                     prompt = prompt.replace(' ',',').replace('　',',').replace(',,',',')
-            
-            filename = uuid.uuid4()
+            if mode == 0:
+                filename = uuid.uuid4()
 
-            payload = {
-                "prompt": f"{prompt} ,high quality, 8k",
-                "steps": 20,
-                "width":512,
-                "height":512,
-                "batch_size": 1,
-                "cfg_scale": 7,
-                "restore_faces": True,
-                "negative_prompt": "low quality,blurry,lowres,duplicate,morbid,deformed,monochrome,greyscale,comic,4koma,2koma,sepia,simple background,rough,unfinished,horror,duplicate legs,duplicate arms,error,worst quality,normal quality",
-            }
-            response = requests.post(url=f'http://127.0.0.1:7861/sdapi/v1/txt2img', json=payload,headers={"Content-Type": "application/json"})
-            r = response.json()
-            with open(f".\\result\\{filename}.png", 'wb') as f:
-                f.write(base64.b64decode(r['images'][0]))
+                payload = {
+                    "prompt": f"{prompt} ,high quality, 8k",
+                    "steps": 20,
+                    "width":512,
+                    "height":512,
+                    "batch_size": 1,
+                    "cfg_scale": 7,
+                    "restore_faces": True,
+                    "negative_prompt": "low quality,blurry,lowres,duplicate,morbid,deformed,monochrome,greyscale,comic,4koma,2koma,sepia,simple background,rough,unfinished,horror,duplicate legs,duplicate arms,error,worst quality,normal quality",
+                }
+                response = requests.post(url=f'http://127.0.0.1:7861/sdapi/v1/txt2img', json=payload,headers={"Content-Type": "application/json"})
+                r = response.json()
+                with open(f".\\result\\{filename}.png", 'wb') as f:
+                    f.write(base64.b64decode(r['images'][0]))
 
-            await msg.delete()
-            await message.reply('完成したよ！',file=discord.File(f'.\\result\\{filename}.png'))
-            os.remove(f".\\result\\{filename}.png")
+                await msg.delete()
+                await message.reply('完成したよ！',file=discord.File(f'.\\result\\{filename}.png'))
+                os.remove(f".\\result\\{filename}.png")
+            else:
+                try:
+                    response = requests.get(url=f'http://192.168.255.186:9876/image?token={message.author.id}&prompt={prompt}&step=50')
+                    await message.reply(file=discord.File(BytesIO(response.content),"result.png"))
+                except Exception as e:
+                    print(e)
+                    await message.reply(f"生成に失敗しました！\n```\n{e}\n```")
         else:
             await message.channel.send('作りたい画像の単語を指定してね！\n（例：@USB無くした かわいい猫）')
